+++ inc/uart.h
@@ -1,5 +1,5 @@
-#ifndef	_MINI_UART_H
-#define	_MINI_UART_H
+#ifndef	_UART_H
+#define	_UART_H
 
 void uart_init(void);
 void uart_send(char c);
@@ -8,4 +8,4 @@ char uart_getc(void);
 void uart_putc(char c);
 void uart_puts(char *s);
 
-#endif// _MINI_UART_H
+#endif//_UART_H
