+++ src/uart.c
@@ -1,0 +1,103 @@
+#include "uart.h"
+#include "gpio.h"
+#include "mailbox.h"
+
+/* PL011 UART registers */
+#define UART0_DR        ((volatile unsigned int*)(MMIO_BASE+0x00201000))
+#define UART0_FR        ((volatile unsigned int*)(MMIO_BASE+0x00201018))
+#define UART0_IBRD      ((volatile unsigned int*)(MMIO_BASE+0x00201024))
+#define UART0_FBRD      ((volatile unsigned int*)(MMIO_BASE+0x00201028))
+#define UART0_LCRH      ((volatile unsigned int*)(MMIO_BASE+0x0020102C))
+#define UART0_CR        ((volatile unsigned int*)(MMIO_BASE+0x00201030))
+#define UART0_IMSC      ((volatile unsigned int*)(MMIO_BASE+0x00201038))
+#define UART0_ICR       ((volatile unsigned int*)(MMIO_BASE+0x00201044))
+
+/**
+ * Set baud rate and characteristics (115200 8N1) and map to GPIO
+ */
+void uart_init()
+{
+    register unsigned int r;
+
+    /* initialize UART */
+    *UART0_CR = 0;         // turn off UART0
+
+    /* set up clock for consistent divisor values */
+    mbox[0] = 9*4;
+    mbox[1] = MBOX_REQUEST_CODE;
+    mbox[2] = MBOX_TAG_SET_CLK_RATE; // set clock rate
+    mbox[3] = 12;
+    mbox[4] = 8;
+    mbox[5] = 2;           // UART clock
+    mbox[6] = 4000000;     // 4Mhz
+    mbox[7] = 0;           // clear turbo
+    mbox[8] = MBOX_END_TAG;
+    mbox_call(MBOX_CH_PROPT_ARM_VC);
+
+    /* map UART0 to GPIO pins */
+    r=*GPFSEL1;
+    r&=~((7<<12)|(7<<15)); // gpio14, gpio15
+    r|=(4<<12)|(4<<15);    // alt0
+    *GPFSEL1 = r;
+    *GPPUD = 0;            // enable pins 14 and 15
+    r=150; while(r--) { asm volatile("nop"); }
+    *GPPUDCLK0 = (1<<14)|(1<<15);
+    r=150; while(r--) { asm volatile("nop"); }
+    *GPPUDCLK0 = 0;        // flush GPIO setup
+
+    *UART0_ICR = 0x7FF;    // clear interrupts
+    *UART0_IBRD = 2;       // 115200 baud
+    *UART0_FBRD = 0xB;
+    *UART0_LCRH = 0b11<<5; // 8n1
+    *UART0_CR = 0x301;     // enable Tx, Rx, FIFO
+}
+
+/**
+ * Send a character by uart
+ */
+void uart_send(char c) {
+    do{/* wait until we can send */
+        asm volatile("nop");
+    }while(*UART0_FR&0x20);
+    /* write the character to the buffer */
+    *UART0_DR=c;
+}
+
+/**
+ * Receivce character by uart
+ */
+char uart_recv() {
+    do{/* wait until something is in the buffer */
+        asm volatile("nop");
+    }while(*UART0_FR&0x10);
+    /* read it and return */
+    return (char)(*UART0_DR);
+}
+
+/**
+ * UART user interface receive character
+ */
+char uart_getc() {
+    char r = uart_recv();
+    /* convert carrige return to newline */
+    return r == '\r' ? '\n' : r;
+}
+
+/**
+ * UART user interface put character
+ */
+void uart_putc(char c) {
+    /* convert newline to carrige return + newline */
+    if(c == '\n')
+        uart_send('\r');
+    uart_send(c);
+}
+
+/**
+ * UART user interface put string
+ */
+void uart_puts(char *s) {
+    while(*s) {
+        uart_putc(*s++);
+    }
+}
